// frontend/src/types/schedule.ts
export type ScheduleDayKey =
  | 'monday'
  | 'tuesday'
  | 'wednesday'
  | 'thursday'
  | 'friday'
  | 'saturday'
  | 'sunday';

export interface ScheduleWeekdayAssignment {
  monday: string | null;
  tuesday: string | null;
  wednesday: string | null;
  thursday: string | null;
  friday: string | null;
  saturday: string | null;
  sunday: string | null;
}

export interface Schedule {
  id: string;
  name: string;
  weekdays: ScheduleWeekdayAssignment;
}